// front/src/app/api-client/model/meal-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Repas planifié
 */
export interface MealDTO { 
    id?: number;
    date?: string;
    creneau?: MealDTO.CreneauEnum;
    portions?: number;
    /**
     * Recette type, absente pour un libellé libre
     */
    recetteId?: number | null;
    /**
     * Libellé libre, absent si une recette est choisie
     */
    libelle?: string | null;
    /**
     * À afficher : nom de la recette ou libellé libre
     */
    titre?: string;
}
export namespace MealDTO {
    export const CreneauEnum = {
        PetitDejeuner: 'PETIT_DEJEUNER',
        Midi: 'MIDI',
        Souper: 'SOUPER',
    } as const;
    export type CreneauEnum = typeof CreneauEnum[keyof typeof CreneauEnum];
}