// front/src/app/api-client/model/recipe-sheet-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipeIngredientDTO } from './recipe-ingredient-dto';


/**
 * Fiche recette : quantités recalculées pour un nombre de portions donné
 */
export interface RecipeSheetDTO { 
    id?: number;
    nom?: string;
    description?: string | null;
    /**
     * Portions de référence de la recette
     */
    portionsRecette?: number;
    /**
     * Portions demandées : les quantités sont calculées pour ce nombre
     */
    portions?: number;
    /**
     * En minutes
     */
    tempsPreparation?: number | null;
    instructions?: string | null;
    /**
     * Quantités au prorata, arrondies à 2 décimales
     */
    ingredients?: Array<RecipeIngredientDTO>;
}